chore:tile_gen: Add fabric_automation.py and move tile_gen there
