+++ FABulous/fabric_generator/fabric_gen.py
@@ -42,7 +42,6 @@ from FABulous.fabric_generator.code_generation_Verilog import VerilogWriter
 from FABulous.fabric_generator.code_generation_VHDL import VHDLWriter
 from FABulous.fabric_generator.code_generator import codeGenerator
 from FABulous.fabric_generator.file_parser import (
-    parseBelFile,
     parseConfigMem,
     parseList,
     parseMatrix,
@@ -2839,130 +2838,3 @@ def generateUserDesignTopWrapper(
     logger.info(
         f"Generated user design top wrapper {output} with {len(top_wrapper)} lines"
     )
-
-
-def generate_custom_tile_config(tile_path: Path) -> Path:
-    """
-    Generates a custom tile configuration for a given tile folder
-    or path to bel folder.
-    A tile .csv file and a switch matrix .list file will be generated.
-
-    The provided path may contain bel files, which will be included
-    in the generated tile .csv file as well as the generated
-    switch matrix .list file.
-
-    Parameters
-    ----------
-
-    tile_path : Path
-        The path to the tile folder. If the path is a file, the parent
-        directory will be used as the tile folder.
-
-    Returns
-    -------
-    Path
-        Path to the generated tile .csv file.
-
-    """
-    tile_name: str = ""
-    project_tile_dir: Path = Path(os.getenv("FAB_PROJ_DIR")).absolute() / "Tile"
-
-    tile_files = {}
-    tile_csv: Path
-    tile_bels: list[Path] = []
-    tile_carrys = []
-    tile_switchmatrix: Path
-    csv_out: list[str] = []
-
-    tile_path = Path(tile_path).absolute()
-
-    logger.info(f"Generating custom tile config {tile_path}")
-
-    if tile_path.is_file():
-        tile_path = tile_path.parent
-
-    tile_name = tile_path.stem
-    tile_csv = tile_path / f"{tile_name}.csv"
-    tile_switchmatrix = tile_path / f"{tile_name}_switch_matrix.list"
-
-    if not tile_path.is_relative_to(project_tile_dir.absolute()):
-        raise ValueError(f"Path {tile_path} is not a valid tile path")
-
-    if not tile_path.exists():
-        tile_path.mkdir()
-    else:
-        tile_files = tile_path.rglob("*")
-
-    for file in tile_files:
-        if not file.is_file():
-            logger.debug(f"Skipping file {file} since it is not a file.")
-            continue
-        if (
-            "configmem" in file.name.lower()
-            or "config_mem" in file.name.lower()
-            or "switchmatrix" in file.name.lower()
-            or "switch_matrix" in file.name.lower()
-        ):
-            logger.debug(
-                f"File {file}is most likely a generated file and will be ignored."
-            )
-            continue
-
-        if file.suffix.lower() in [".vhdl", ".vhd", ".v", ".sv"]:
-            logger.info(f"Found BEL file {file} for custom tile {tile_name}")
-            tile_bels.append(file)
-
-        elif file.suffix.lower() == ".csv":
-            logger.warning(
-                f"Found tile config CSV file {file} for custom tile {tile_name}, nothing to do here."
-            )
-            return file
-        elif file.suffix.lower() == ".list":
-            tile_switchmatrix = file
-            logger.warning(
-                f"Found tile tile_switchmatrix list file {file} for custom tile {tile_name}, no switchmatrix list file will be generated."
-            )
-        else:
-            logger.warning(
-                f"File {file} in custom tile {tile_name} is not a valid config or bel file."
-            )
-
-    has_reset = False
-    has_enable = False
-    for file in tile_bels:
-        if file.suffix.lower() in [".v", ".sv"]:
-            bel = parseBelFile(file, "", "verilog")
-        else:
-            bel = parseBelFile(file, "", "vhdl")
-        if "RESET" in bel.localShared.keys():
-            has_reset = True
-        if "ENABLE" in bel.localShared.keys():
-            has_enable = True
-        for carry in bel.carry:
-            if carry not in tile_carrys:
-                tile_carrys.append(carry)
-    # Create tile config CSV file
-    logger.info(f"Creating tile config CSV file {tile_csv}")
-    tile_csv.touch()
-
-    csv_out.append(f"TILE,{tile_name}")
-    csv_out.append(f"INCLUDE,./../include/Base.csv")
-    for i, carry in enumerate(tile_carrys):
-        csv_out.append(f'NORTH,Co{i},0,-1,Ci{i},1,CARRY="{carry}"')
-    if has_reset:
-        csv_out.append("JUMP,J_SRST_BEG,0,0,J_SRST_END,1,SHARED_RESET")
-    if has_enable:
-        csv_out.append("JUMP,J_SEN_BEG,0,0,J_SEN_END,1,SHARED_ENABLE")
-    for bel in tile_bels:
-        csv_out.append(f"BEL,./{bel.relative_to(tile_path)}")
-    if tile_switchmatrix.exists():
-        csv_out.append(f"MATRIX,{tile_switchmatrix.relative_to(tile_path)}")
-    else:
-        csv_out.append("MATRIX,GENERATE")
-    csv_out.append("EndTILE")
-
-    with tile_csv.open("w", encoding="utf-8") as file:
-        file.write("\n".join(csv_out))
-
-    return tile_csv
-
